requirements.txt
# 核心依赖
qwen-agent>=0.0.8
python-dotenv>=1.0.0

# Web界面支持
streamlit>=1.28.0
gradio>=4.0.0

# 可选依赖（用于增强功能）
requests>=2.31.0
aiohttp>=3.8.0

# 开发依赖（可选）
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
